rename ping command file (camelCase) and use constant messages, description object
